mqtt_toolbox/requirements.txt
paho-mqtt
python-dotenv
numpy
